// src/types/enums/UPermissions.ts
export enum UPermission {
    CREATE_OLYMPIAD = "CREATE_OLYMPIAD", 
    CONFIGURE_AREAS = "CONFIGURE_AREAS", 
    REGISTER_ACADEMIC_RESPONSIBLE = "REGISTER_ACADEMIC_RESPONSIBLE", 
    REGISTER_EVALUATOR = "REGISTER_EVALUATOR", 
    REGISTER_COMPETITOR = "REGISTER_COMPETITOR", 
    GRADE_COMPETITOR = "GRADE_COMPETITOR",
    EDIT_SCORE_CUT = "EDIT_SCORE_CUT",
    FILTER_COMPETITOR_BY_AREA = "FILTER_COMPETITOR_BY_AREA",
    VIEW_OLYMPIAD = "VIEW_OLYMPIAD",
}